LPD-56911 Move components to a folder
