+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main/components/modal/MultipleFilesUploadModalContent.tsx
@@ -9,7 +9,7 @@ import {sub} from 'frontend-js-web';
 import React from 'react';
 
 import {AssetLibrary} from '../../../types/AssetLibrary';
-import MultipleFileUploader from '../MultipleFileUploader';
+import MultipleFileUploader from '../multiple_file_uploader/MultipleFileUploader';
 
 export default function MultipleFilesUploadModalContent({
 	assetLibraries,
