+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main/components/multiple_file_uploader/DragZoneBackground.tsx
@@ -6,7 +6,7 @@
 import ClayButton from '@clayui/button';
 import React from 'react';
 
-import {getImage} from '../util/getImage';
+import {getImage} from '../../util/getImage';
 
 export default function DragZoneBackground() {
 	return (
